Lift context menu to tab scope
